requirements.txt
discord-py-slash-command==4.2.1
psycopg==3.2.2
py-cord==2.6.1
python-dotenv==1.0.1
